requirements.txt
Flask==2.0.2
pdf2image==1.16.0
pytesseract==0.3.8
Pillow==8.4.0